xinya_order_platform_v4_fixed11_project/requirements.txt
streamlit==1.37.0
pandas==2.2.2
python-dotenv==1.0.1
Pillow==10.4.0
reportlab==4.2.2
openpyxl==3.1.2